requirements.txt
Markdown==2.5.2
argparse==1.2.1
django-debug-toolbar==1.3.2
django-tagging==0.3.2
docutils==0.12
pytz==2014.4
sqlparse==0.1.15
tagging==0.2.1
wsgiref==0.1.2
